+++ src/dihedral.h
@@ -43,7 +43,7 @@ class Dihedral : protected Pointers {
   // KOKKOS host/device flag and data masks
 
   ExecutionSpace execution_space;
-  uint64_t datamask_read, datamask_modify;
+  unsigned int datamask_read, datamask_modify;
   int copymode, kokkosable;
 
   Dihedral(class LAMMPS *);
